intelligence-silo/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "intelligence-silo"
version = "0.1.0"
description = "Neural intelligence silo — society of minds architecture with hierarchical memory and small language model matrix"
requires-python = ">=3.11"
dependencies = [
    # Core ML
    "torch>=2.2.0",
    "transformers>=4.38.0",
    "sentence-transformers>=2.5.0",
    "safetensors>=0.4.0",

    # Vector memory
    "faiss-cpu>=1.7.4",
    "numpy>=1.26.0",

    # Google services
    "google-cloud-storage>=2.14.0",
    "google-cloud-aiplatform>=1.43.0",
    "google-auth>=2.27.0",
    "firebase-admin>=6.4.0",

    # Decision engine integration
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.6.0",
    "pyyaml>=6.0.1",
    "httpx>=0.27.0",

    # Packaging & distribution
    "pyinstaller>=6.4.0",
    "cryptography>=42.0.0",

    # Utilities
    "anthropic>=0.18.0",
    "rich>=13.7.0",
    "typer>=0.9.0",
    "msgpack>=1.0.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.2.0",
]
gpu = [
    "torch>=2.2.0+cu121",
    "faiss-gpu>=1.7.4",
]

[project.scripts]
intel-silo = "intelligence_silo.cli:app"

[tool.setuptools.packages.find]
where = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"
